--- src/components/implementation/no_interface/ekf/gpsdata.h
@@ -1,29 +1,0 @@
-#ifndef GPSDATA_H
-#define GPSDATA_H
-
-double input[25][16] = {{-11602023.9489137008786201,14063117.4931115992367268,18811434.3112745992839336,-20853271.5736341997981071,1806977.2118581600952893,16542682.1237923000007868,-14355926.0172339994460344,8650961.8841098193079233,20736354.9805863983929157,7475239.6753052901476622,12966181.2771377004683018,21931576.7921750992536545,23568206.4173782989382744,26183921.4577450007200241,24652215.2627705000340939,25606982.9330466017127037},{-11602700.4096150007098913,14060708.1637619994580746,18812823.4023028016090393,-20855049.9291185997426510,1805887.1306580700911582,16540582.4659656994044781,-14356344.1729805991053581,8648384.4768619798123837,20737164.3397033996880054,7472917.3215693095698953,12967714.4596338998526335,21931442.6029887981712818,23568427.7909861989319324,26184404.1127416007220745,24652621.9011856988072395,25606499.4748000986874104},
-{-11603377.0261803008615971,14058298.6961425002664328,18814212.0761809013783932,-20856828.1167653985321522,1804797.2804981300141662,16538482.4609003998339176,-14356762.4791433997452259,8645806.9947465099394321,20737973.2627678997814655,7470595.0720982002094388,12969247.7736987993121147,21931307.9468086995184422,23568650.0894557014107704,26184884.7086125016212463,24653025.2764103002846241,25606016.6971120014786720},
-{-11604053.7986267991364002,14055889.0902858991175890,18815600.3328956998884678,-20858606.1364935003221035,1803707.6613832199946046,16536382.1086646001785994,-14357180.9357223007827997,8643229.4378561992198229,20738781.7497542984783649,7468272.9269468197599053,12970781.2192928008735180,21931172.8236371017992496,23568869.5260895006358624,26185366.6481501981616020,24653428.8435873985290527,25605534.4603805989027023},
-{-11604730.7269448004662991,14053479.3463228996843100,18816988.1723781004548073,-20860383.9882667995989323,1802618.2732910700142384,16534281.4092740993946791,-14357599.5427193008363247,8640651.8062730506062508,20739589.8006406985223293,7465950.8861416298896074,12972314.7963951993733644,21931037.2334740012884140,23569094.4420915991067886,26185845.7782028988003731,24653834.8537949994206429,25605048.9604584984481335},
-{-11605407.8111640997231007,14051069.4642411991953850,18818375.5946411006152630,-20862161.6719904989004135,1801529.1162350000813603,16532180.3628132995218039,-14358018.3001422006636858,8638074.1000416092574596,20740397.4154165014624596,7463628.9497939096763730,12973848.5049293003976345,21930901.1763249002397060,23569315.4143446013331413,26186327.8049918003380299,24654241.1781065985560417,25604567.3344845995306969},
-{-11606085.0512815993279219,14048659.4441480003297329,18819762.5996288992464542,-20863939.1876304000616074,1800440.1901911599561572,16530078.9692954998463392,-14358437.2079997994005680,8635496.3192011900246143,20741204.5940730981528759,7461307.1179005000740290,12975382.3448939993977547,21930764.6521883010864258,23569537.8873163014650345,26186808.2263608016073704,24654645.1117384992539883,25604081.9392636008560658},
-{-11606762.4473100006580353,14046249.2860924992710352,18821149.1873193010687828,-20865716.5351110994815826,1799351.4951612299773842,16527977.2287824992090464,-14358856.2662888001650572,8632918.4638650994747877,20742011.3365787006914616,7458985.3905708203092217,12976916.3162136003375053,21930627.6610695011913776,23569760.0636343993246555,26187289.5027904994785786,24655052.4830632992088795,25603599.6850817985832691},
-{-11607439.9992581997066736,14043838.9901378005743027,18822535.3576819002628326,-20867493.7143885008990765,1798263.0311275599524379,16525875.1412992998957634,-14359275.4750231001526117,8630340.5340407807379961,20742817.6429346017539501,7456663.7678293604403734,12978450.4188688006252050,21930490.2029686011373997,23569981.9083983004093170,26187768.8422459997236729,24655456.8704008981585503,25603116.4885880984365940},
-{-11608117.7071295995265245,14041428.5563670992851257,18823921.1106748990714550,-20869270.7253601998090744,1797174.7981080398894846,16523772.7069393992424011,-14359694.8342019002884626,8627762.5298271290957928,20743623.5131133012473583,7454342.2497338298708200,12979984.6528181005269289,21930352.2778877988457680,23570205.8646384999155998,26188253.1899140998721123,24655862.4792538993060589,25602632.6115358993411064},
-{-11608795.5709421001374722,14039017.9848111998289824,18825306.4462864995002747,-20871047.5680137984454632,1796086.7960656299255788,16521669.9256903994828463,-14360114.3438322991132736,8625184.4512723106890917,20744428.9471033997833729,7452020.8363409303128719,12981519.0180207993835211,21930213.8858294002711773,23570427.8664544001221657,26188734.3965431004762650,24656267.6169511005282402,25602148.1411762982606888},
-{-11609473.5906990002840757,14036607.2755532003939152,18826691.3644751012325287,-20872824.2422708012163639,1794999.0250037000514567,16519566.7976179998368025,-14360534.0039195995777845,8622606.2984361592680216,20745233.9448899999260902,7449699.5277419695630670,12983053.5144130997359753,21930075.0267974995076656,23570650.3219759985804558,26189215.4635703004896641,24656671.8995876014232635,25601667.6320159994065762},
-{-11610151.7664094995707273,14034196.4286551997065544,18828075.8652108013629913,-20874600.7480676993727684,1793911.4849164399784058,16517463.3227698002010584,-14360953.8144662007689476,8620028.0713985096663237,20746038.5064514987170696,7447378.3239304702728987,12984588.1419961005449295,21929935.7007904984056950,23570873.1090517006814480,26189696.8272514007985592,24657077.3339385986328125,25601183.0395046994090080},
-{-11610830.0980858001857996,14031785.4441682007163763,18829459.9484703987836838,-20876377.0853387005627155,1792824.1757994499057531,16515359.5011966004967690,-14361373.7754756007343531,8617449.7702285703271627,20746842.6317701004445553,7445057.2250017002224922,12986122.9007034003734589,21929795.9078129008412361,23571094.6397118009626865,26190179.3251966014504433,24657484.6529132016003132,25600699.4416556991636753},
-{-11611508.5857305992394686,14029374.3221777006983757,18830843.6142107993364334,-20878153.2540492005646229,1791737.0976289999671280,16513255.3329117000102997,-14361793.8869582004845142,8614871.3949565794318914,20747646.3208399005234241,7442736.2310290103778243,12987657.7904831003397703,21929655.6478679999709129,23571317.6536404006183147,26190658.5076005011796951,24657890.0872642993927002,25600219.0895471982657909},
-{-11612187.2293522991240025,14026963.0627482999116182,18832226.8624009005725384,-20879929.2540921010077000,1790650.2504261699505150,16511150.8180149998515844,-14362214.1489185001701117,8612292.9456460792571306,20748449.5736446008086205,7440415.3420168599113822,12989192.8113291002810001,21929514.9209545999765396,23571542.2729890011250973,26191142.2270611003041267,24658293.6893425993621349,25599735.3346460983157158},
-{-11612866.0289661008864641,14024551.6659204997122288,18833609.6930234991014004,-20881705.0854518003761768,1789563.6341555600520223,16509045.9564973991364241,-14362634.5613554008305073,8609714.4223972298204899,20749252.3901568017899990,7438094.5580563498660922,12990727.9631776995956898,21929373.7270772010087967,23571765.6359219998121262,26191622.8229327984154224,24658699.8217025995254517,25599252.7314594015479088},
-{-11613544.9845812991261482,14022140.1317561008036137,18834992.1060491017997265,-20883480.7480603009462357,1788477.2488142400979996,16506940.7484122999012470,-14363055.1242757998406887,8607135.8252597600221634,20750054.7703643999993801,7435773.8791862595826387,12992263.2459998000413179,21929232.0662368983030319,23571987.5330366007983685,26192101.5167306996881962,24659106.9487251006066799,25598769.0638093985617161},
-{-11614224.0961968991905451,14019728.4603524003177881,18836374.1014277003705502,-20885256.2418182007968426,1787391.0944181799422950,16504835.1938502006232738,-14363475.8376898001879454,8604557.1542639899998903,20750856.7142618000507355,7433453.3054846199229360,12993798.6597405001521111,21929089.9384372010827065,23572212.1698355004191399,26192584.8348365016281605,24659511.3186132013797760,25598287.1935316994786263},
-{-11614903.3638345003128052,14017316.6517277993261814,18837755.6791551001369953,-20887031.5667084008455276,1786305.1709331099409610,16502729.2928037997335196,-14363896.7015940994024277,8601978.4095227196812630,20751658.2218171991407871,7431132.8369907401502132,12995334.2043703999370337,21928947.3436792008578777,23572433.9098982997238636,26193065.3609073981642723,24659918.7073891013860703,25597804.9916997998952866},
-{-11615582.7874893993139267,14014904.7059931997209787,18839136.8391736000776291,-20888806.7226402983069420,1785219.4783696499653161,16500623.0453533995896578,-14364317.7159989997744560,8599399.5910640191286802,20752459.2930257990956306,7428812.4737586695700884,12996869.8798501994460821,21928804.2819650992751122,23572658.6513984985649586,26193548.1555067002773285,24660325.0840523988008499,25597322.2140106000006199},
-{-11616262.3671784996986389,14012492.6231826990842819,18840517.5814694985747337,-20890581.7095695994794369,1784134.0167101800907403,16498516.4515240993350744,-14364738.8809060007333755,8596820.6989716701209545,20753259.9278649017214775,7426492.2158625600859523,12998405.6861275006085634,21928660.7532982006669044,23572882.7297905012965202,26194030.4265995994210243,24660732.8916335999965668,25596841.2162435986101627},
-{-11616942.1029125005006790,14010080.4033525995910168,18841897.9060166999697685,-20892356.5274465009570122,1783048.7859403199981898,16496409.5113475006073713,-14365160.1963243000209332,8594241.7332799397408962,20754060.1263274997472763,7424172.0633234297856688,12999941.6231851000338793,21928516.7576785981655121,23573105.2551130987703800,26194510.3070125989615917,24661138.8145914003252983,25596357.5136927999556065},
-{-11617621.9946924000978470,14007668.0465942993760109,18843277.8127687983214855,-20894131.1761795990169048,1781963.7860713400878012,16494302.2249049004167318,-14365581.6622592005878687,8591662.6940501499921083,20754859.8883983008563519,7421852.0162422796711326,13001477.6909523997455835,21928372.2951112985610962,23573329.6650592982769012,26194992.9794606007635593,24661542.6609732992947102,25595876.9347308985888958},
-{-11618302.0425349995493889,14005255.5529417991638184,18844657.3017124012112617,-20895905.6557381004095078,1780879.0170771600678563,16492194.5922053009271622,-14366003.2787072006613016,8589083.5813942104578018,20755659.2140460014343262,7419532.0746282795444131,13003013.8894202001392841,21928227.3655956983566284,23573552.3125334009528160,26195473.3659299984574318,24661950.1370006017386913,25595393.4415826015174389}};
-
-#endif /* GPSDATA_H */
