+++ src/components/implementation/no_interface/llbooter/Makefile
@@ -4,7 +4,7 @@ COMPONENT=llboot_comp.o
 INTERFACES=
 DEPENDENCIES=
 IF_LIB=
-ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_defkernel_api
+ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_defkernel_api -lekf
 
 include ../../Makefile.subsubdir
 MANDITORY_LIB=simple_stklib.o
