+++ src/components/lib/tiny_ekf/gps_ekf.c
@@ -205,7 +205,7 @@ int main(int argc, char ** argv)
     for (j=0; j<25; ++j) {
         //fprintf(ofp, "%f,%f,%f\n", 
                 //Pos_KF[j][0]-mean_Pos_KF[0], Pos_KF[j][1]-mean_Pos_KF[1], Pos_KF[j][2]-mean_Pos_KF[2]);
-        PRINTC("%f %f %f\n", Pos_KF[j][0], Pos_KF[j][1], Pos_KF[j][2]);
+        //PRINTC("%f %f %f\n", Pos_KF[j][0], Pos_KF[j][1], Pos_KF[j][2]);
     }
     
     // Done!
